+++ smtpdfix/__init__.py
@@ -6,7 +6,7 @@ __all__ = (
     "smtpd",
     "SMTPDFix",
 )
-__version__ = "0.5.3"
+__version__ = "0.6.0.dev"
 
 from .authenticator import Authenticator
 from .configuration import Config